refactor(2024/01): move extension to separate file; rename input file
Part 2 uses the same input file. Rename to reflect this.
