+++ solutions/2024/day01/kotlin/src/main/kotlin/se/fzy/adventofcode/Part1.kt
@@ -3,7 +3,7 @@ package se.fzy.adventofcode
 import kotlin.math.absoluteValue
 
 fun main() {
-    val (a, b) = Resources.readInput("input_part1.txt")
+    val (a, b) = Resources.readInput("puzzle_input.txt")
         .lineSequence()
         .filter { it.isNotBlank() }
         .map { line -> line.split("""\s+""".toRegex()).map(String::toInt) }
@@ -14,15 +14,4 @@ fun main() {
 
     val distance = a.zip(b).sumOf { (a, b) -> (a - b).absoluteValue }
     println(distance)
-}
-
-fun <T> List<List<T>>.transpose(): Sequence<List<T>> {
-    if (isEmpty()) return emptySequence()
-    require(all { it.size == first().size }) { "Cannot transpose lists of non-uniform size" }
-
-    return sequence {
-        first().indices.forEach { i ->
-            yield(this@transpose.indices.map { j -> get(j)[i] }.toList())
-        }
-    }
 }
