+++ solutions/2024/day01/kotlin/src/main/resources/puzzle_input.txt
